% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gintreg.R
\name{gintreg}
\alias{gintreg}
\title{Fit a generalized normal interval regression model to interval data}
\usage{
gintreg(location, scale, data, start, weights, thresholds)
}
\arguments{
\item{location}{an object of class "formula"
that specifies the equation for the conditional mean.}

\item{scale}{an object of class "formula"
that specifies the equation for the conditional standard deviation.}

\item{data}{a data.frame containing the variables in the model.}

\item{start}{an optional vector of starting values for the parameters.}

\item{weights}{an optional vector of weights.}

\item{thresholds}{a vector of J - 1 thresholds when there are J intervals.}
}
\value{
an object of class "gintreg"
}
\description{
\code{gintreg} fits a generalized normal interval regression model,
i.e., a normal interval regression model with conditional heteroskedasticity,
to interval data.
The dependent variable must be a factor representing J intervals,
and the user must provide J - 1 thresholds.
}
