% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intreg.R
\name{intreg}
\alias{intreg}
\title{Fit a normal interval regression model to interval data}
\usage{
intreg(formula, data, start, weights, thresholds)
}
\arguments{
\item{formula}{an object of class "formula".}

\item{data}{a data.frame containing the variables in the model.}

\item{start}{an optional vector of starting values for the parameters.}

\item{weights}{an optional vector of weights.}

\item{thresholds}{a vector of J thresholds when there are J + 1 intervals.}
}
\value{
an object of class "intreg"
}
\description{
\code{intreg} fits a normal interval regression model to interval data.
The dependent variable must be a factor representing J + 1 intervals,
and the user must provide J thresholds.
}
\examples{
intreg(
  q2 ~ sex,
  data       = ias2009febmay,
  thresholds = c(-.5, .5, 1:5)
)
}
