% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidier.R
\name{tidy.hintreg}
\alias{tidy.hintreg}
\title{Tidy an hintreg object}
\usage{
\method{tidy}{hintreg}(lFit, conf.int = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{conf.int}{logical indicating whether or not to include a confidence interval in the tidied output.}

\item{conf.level}{the confidence level to use for the confidence interval if conf.int = TRUE.}

\item{...}{additional arguments not used, but needed to match generic signature only.}

\item{x}{an object of class "hintreg".}
}
\value{
a tidy \code{\link[tibble:tibble]{tibble::tibble()}} summarizing component-level information about the model.
}
\description{
Tidy an hintreg object
}
\examples{
lout <- hintreg(
  q2 ~ sex,
     ~ sex,
     ~ sex,
     ~ sex,
  data       = ias2009febmay,
  thresholds = c(-.5, .5, 1:5)
)
tidy(lout)
}
