% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intreg.R
\name{tidy.intreg}
\alias{tidy.intreg}
\title{Tidy an intreg object}
\usage{
tidy.intreg(lFit, conf.int = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{lFit}{an object of class "intreg".}

\item{conf.int}{logical indicating whether or not to include a confidence interval in the tidied output.}

\item{conf.level}{the confidence level to use for the confidence interval if conf.int = TRUE.}

\item{...}{additional arguments not used, but needed to match generic signature only.}
}
\value{
a tidy \code{\link[tibble:tibble]{tibble::tibble()}} summarizing component-level information about the model.
}
\description{
Tidy an intreg object
}
\examples{
lout <- intreg(
  q2 ~ sex,
  data       = ias2009febmay,
  thresholds = c(-.5, .5, 1, 2, 3, 4, 5)
)
tidy(lout)
}
