% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hintreg.R
\name{hintreg}
\alias{hintreg}
\title{Fit a heterogeneous normal interval regression model to interval data}
\usage{
hintreg(
  location,
  scale,
  lower,
  upper,
  data,
  start,
  weights,
  threshbelow,
  threshabove,
  limenlb,
  limenub
)
}
\arguments{
\item{location}{an object of class "formula"
that specifies the equation for the conditional mean.}

\item{scale}{an object of class "formula"
that specifies the equation for the conditional standard deviation.}

\item{lower}{an object of class "formula"
that specifies the equation for the lower bound of the indifference limen.}

\item{upper}{an object of class "formula"
that specifies the equation for the upper bound of the indifference limen.}

\item{data}{a data.frame containing the variables in the model.}

\item{start}{an optional vector of starting values for the parameters.}

\item{weights}{an optional vector of weights.}

\item{threshbelow}{a vector of thresholds below 0}

\item{threshabove}{a vector of thresholds above 0}

\item{limenlb}{numeric.
A lower bound on the heterogeneous indifference limen not smaller than max(threshbelow).}

\item{limenub}{numeric.
An upper bound on the heterogeneous indifference limen not greater than min(threshabove).}
}
\value{
an object of class "hintreg"
}
\description{
\code{hintreg} fits a heterogeneous normal interval regression model
to interval data with an indifference limen,
i.e., a normal interval regression model with conditional heteroskedasticity
and a heterogeneous indifference limen.
The dependent variable must be a factor representing J + 1 intervals.
The user must provide thresholds below and above 0, respectively,
expect for an indifference limen (so there must be J - 2 known thresholds),
and lower and upper bounds on the heterogeneous indifference limen.
}
